Broke up `ToggleInternal()`, fixed label positioning.
It was large, so I split it into logical chunks. Still a bit messy but easier to follow flow.

A11y labels now follow the same offset logic as the widget label, so they are properly placed. As well, the widget label will now scoot over should the knob offset move the knob outside of the toggle frame.